api/requirements.txt
fastapi
fastapi_utils
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
loguru
pydantic
pydantic[email]
typing
itsdangerous
emails
pika
pysftp
psycopg2
orjson